# docs/contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bookstore Microservices API",
    "description": "Complete API specification for the Bookstore microservices platform including Catalog, Order, and User services accessible through the API Gateway",
    "version": "1.0.0",
    "contact": {
      "name": "Bookstore API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "API Gateway - Local Development"
    }
  ],
...
  "paths": {
...
    "/api/v1/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get all orders",
        "description": "Retrieves all orders in the system",
        "operationId": "getAllOrders",
        "responses": {
          "200": {
            "description": "Orders retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Orders"],
        "summary": "Create a new order",
        "description": "Creates a new order and validates book availability",
        "operationId": "createOrder",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid order data or insufficient stock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get order by ID",
        "description": "Retrieves order details by order ID",
        "operationId": "getOrderById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the order to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Orders"],
        "summary": "Delete order",
        "description": "Deletes an order by ID",
        "operationId": "deleteOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the order to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Order deleted successfully"
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/user/{userId}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get user orders",
        "description": "Retrieves all orders for a specific user",
        "operationId": "getUserOrders",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "ID of the user",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User orders retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/status/{status}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get orders by status",
        "description": "Retrieves all orders with a specific status",
        "operationId": "getOrdersByStatus",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "required": true,
            "description": "Order status to filter by",
            "schema": {
              "type": "string",
              "enum": ["PENDING", "CONFIRMED", "SHIPPED", "DELIVERED", "CANCELLED"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Orders retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{id}/status": {
      "put": {
        "tags": ["Orders"],
        "summary": "Update order status",
        "description": "Updates the status of an existing order",
        "operationId": "updateOrderStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the order to update",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "description": "New status for the order",
            "schema": {
              "type": "string",
              "enum": ["PENDING", "CONFIRMED", "SHIPPED", "DELIVERED", "CANCELLED"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication token. Include in Authorization header as 'Bearer {token}'"
      }
    },
    "schemas": {
...
      "OrderDto": {
        "type": "object",
        "required": ["userId", "items"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the order (MongoDB ObjectId)",
            "readOnly": true,
            "example": "507f1f77bcf86cd799439011"
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "description": "Identifier of the user who placed the order",
            "example": 1
          },
          "items": {
            "type": "array",
            "minItems": 1,
            "description": "List of items in the order",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          },
          "totalAmount": {
            "type": "number",
            "format": "decimal",
            "description": "Total amount for the entire order (auto-calculated)",
            "readOnly": true,
            "example": 45.97
          },
          "status": {
            "type": "string",
            "enum": ["PENDING", "CONFIRMED", "SHIPPED", "DELIVERED", "CANCELLED"],
            "description": "Current status of the order",
            "example": "PENDING"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the order was created",
            "readOnly": true,
            "example": "2026-01-05T21:13:00Z"
          }
        }
      },
      "OrderItemDto": {
        "type": "object",
        "required": ["bookId", "quantity"],
        "properties": {
          "bookId": {
            "type": "integer",
            "format": "int64",
            "description": "Identifier of the book being ordered",
            "example": 1
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "description": "Quantity of books ordered",
            "example": 2
          },
          "price": {
            "type": "number",
            "format": "decimal",
            "description": "Price per unit at the time of order (auto-populated)",
            "readOnly": true,
            "example": 12.99
          },
          "bookTitle": {
            "type": "string",
            "description": "Title of the book (captured at order time, auto-populated)",
            "readOnly": true,
            "example": "The Great Gatsby"
          },
          "bookIsbn": {
            "type": "string",
            "description": "ISBN of the book (captured at order time, auto-populated)",
            "readOnly": true,
            "example": "9781234567890"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the error occurred",
            "example": "2026-01-05T21:13:00"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 404
          },
          "error": {
            "type": "string",
            "description": "Short error type description",
            "example": "Not Found"
          },
          "message": {
            "type": "string",
            "description": "Detailed error message explaining what went wrong",
            "example": "Book with ID 999 not found"
          },
          "path": {
            "type": "string",
            "description": "The request path that caused the error",
            "example": "/api/v1/books/999"
          }
        }
      }
    }
  }
}